frontend: Split `services.ts` into separate files
