+++ frontend/src/common/services/hitasApi/index.ts
@@ -1,0 +1,5 @@
+export * from "./apartment";
+export * from "./codes";
+export * from "./functions";
+export * from "./housingCompany";
+export * from "./reports";
