+++ frontend/src/common/services/index.ts
@@ -1,1 +1,4 @@
-export * from "./services";
+export * from "./apis";
+export * from "./authApi";
+export * from "./hitasApi";
+export * from "./utils";
